+++ db/facelog-db-base/src/main/java/net/gdface/facelog/db/StoreBean.java
@@ -46,11 +46,11 @@ public final class StoreBean
     /** flag whether {@code this} can be modified */
     private Boolean immutable;
     /** columns modified flag */
-    @ApiModelProperty(value="columns modified flag",dataType="long",required=true)
-    private long modified;
+    @ApiModelProperty(value="columns modified flag",dataType="int",required=true)
+    private int modified;
     /** columns initialized flag */
-    @ApiModelProperty(value="columns initialized flag",dataType="long",required=true)
-    private long initialized;
+    @ApiModelProperty(value="columns initialized flag",dataType="int",required=true)
+    private int initialized;
     /** new record flag  */
     @ApiModelProperty(value="new record flag",dataType="boolean",required=true)
     private boolean isNew;        
@@ -118,7 +118,7 @@ public final class StoreBean
      * @return the modified status of columns
      */
     @ThriftField(value=2,requiredness=Requiredness.REQUIRED)
-    public long getModified(){
+    public int getModified(){
         return modified;
     }
 
@@ -126,14 +126,14 @@ public final class StoreBean
      * @param modified the modified status bit to be assigned to {@link #modified}
      */
     @ThriftField()
-    public void setModified(long modified){
+    public void setModified(int modified){
         this.modified = modified;
     }
     /**
      * @return the initialized status of columns
      */
     @ThriftField(value=3,requiredness=Requiredness.REQUIRED)
-    public long getInitialized(){
+    public int getInitialized(){
         return initialized;
     }
 
@@ -141,7 +141,7 @@ public final class StoreBean
      * @param initialized the initialized status bit to be assigned to {@link #initialized}
      */
     @ThriftField()
-    public void setInitialized(long initialized){
+    public void setInitialized(int initialized){
         this.initialized = initialized;
     }
     protected static final <T extends Comparable<T>>boolean equals(T a, T b) {
@@ -411,7 +411,7 @@ public final class StoreBean
     public void resetIsModified()
     {
         checkMutable();
-        modified = 0L;
+        modified = 0;
     }
 
     @Override
@@ -432,7 +432,7 @@ public final class StoreBean
      */
     private void resetInitialized()
     {
-        initialized = 0L;
+        initialized = 0;
     }
     /** reset all fields to initial value, equal to a new bean */
     public void reset(){
@@ -441,8 +441,8 @@ public final class StoreBean
         this.encoding = null;
         this.data = null;
         this.isNew = true;
-        this.modified = 0L;
-        this.initialized = 0L;
+        this.modified = 0;
+        this.initialized = 0;
     }
     @Override
     public boolean equals(Object object)
@@ -590,7 +590,7 @@ public final class StoreBean
      * @see #asNULL()
      */
     public boolean checkNULL(){
-        return 0L == getInitialized();
+        return 0 == getInitialized();
     }
     /** 
      * @param source source list
