+++ db/facelog-db-orm/src/main/java/net/gdface/facelog/dborm/device/FlDeviceGroupBean.java
@@ -60,9 +60,9 @@ public  class FlDeviceGroupBean
     /** flag whether {@code this} can be modified */
     private Boolean immutable;
     /** columns modified flag */
-    private long modified;
+    private int modified;
     /** columns initialized flag */
-    private long initialized;
+    private int initialized;
     /** new record flag  */
     private boolean isNew;        
     /** 
@@ -126,27 +126,27 @@ public  class FlDeviceGroupBean
     /**
      * @return the modified status of columns
      */
-    public long getModified(){
+    public int getModified(){
         return modified;
     }
 
     /**
      * @param modified the modified status bit to be assigned to {@link #modified}
      */
-    public void setModified(long modified){
+    public void setModified(int modified){
         this.modified = modified;
     }
     /**
      * @return the initialized status of columns
      */
-    public long getInitialized(){
+    public int getInitialized(){
         return initialized;
     }
 
     /**
      * @param initialized the initialized status bit to be assigned to {@link #initialized}
      */
-    public void setInitialized(long initialized){
+    public void setInitialized(int initialized){
         this.initialized = initialized;
     }
     protected static final <T extends Comparable<T>>boolean equals(T a, T b) {
@@ -849,7 +849,7 @@ public  class FlDeviceGroupBean
     public void resetIsModified()
     {
         checkMutable();
-        modified = 0L;
+        modified = 0;
     }
 
     @Override
@@ -877,7 +877,7 @@ public  class FlDeviceGroupBean
      */
     private void resetInitialized()
     {
-        initialized = 0L;
+        initialized = 0;
     }
     /** reset all fields to initial value, equal to a new bean */
     public void reset(){
@@ -895,8 +895,8 @@ public  class FlDeviceGroupBean
         /* DEFAULT:'CURRENT_TIMESTAMP'*/
         this.updateTime = null;
         this.isNew = true;
-        this.modified = 0L;
-        this.initialized = 0L;
+        this.modified = 0;
+        this.initialized = 0;
     }
     @Override
     public boolean equals(Object object)
@@ -1128,7 +1128,7 @@ public  class FlDeviceGroupBean
      * @see #asNULL()
      */
     public boolean checkNULL(){
-        return 0L == getInitialized();
+        return 0 == getInitialized();
     }
     /** 
      * @param source source list
