+++ db/facelog-db-orm/src/main/java/net/gdface/facelog/dborm/device/FlDeviceGroupManager.java
@@ -278,7 +278,7 @@ public class FlDeviceGroupManager extends TableManager.BaseAdapter<FlDeviceGroup
         if(null == bean  || null == bean.getId()){
             return false;
         }
-        long modified = bean.getModified();
+        int modified = bean.getModified();
         try{
             bean.resetModifiedExceptPrimaryKeys();
             return 1 == countUsingTemplate(bean);
