+++ db/facelog-db-orm/src/main/java/net/gdface/facelog/dborm/image/FlImageManager.java
@@ -277,7 +277,7 @@ public class FlImageManager extends TableManager.BaseAdapter<FlImageBean>
         if(null == bean  || null == bean.getMd5()){
             return false;
         }
-        long modified = bean.getModified();
+        int modified = bean.getModified();
         try{
             bean.resetModifiedExceptPrimaryKeys();
             return 1 == countUsingTemplate(bean);
